% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{clear_recordings}
\alias{clear_recordings}
\alias{del_recording}
\title{Delete one or all HTTP request recordings}
\usage{
clear_recordings()

del_recording(id)
}
\arguments{
\item{id}{Id of the recording to delete.}
}
\description{
\code{del_recording} deletes a single recording. \code{clear_recordings} deletes
all of them in the current data store.
}
\details{
They both use the data store in the \code{HTTRMOCK_STORE} environment
variable, or the default one if it is not set.
}
\seealso{
Other HTTP mocking: \code{\link{httrmock}},
  \code{\link{list_recordings}},
  \code{\link{mocking_status}},
  \code{\link{stop_recording}},
  \code{\link{with_httrmock_context}}
}

