% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{ct}
\alias{ct}
\alias{setwt}
\title{Change the current working test}
\usage{
ct(x = "")

setwt(x = "")
}
\arguments{
\item{x}{Sub-test to step into. It might be
\itemize{
\item the (unique) beginning of the name of a test file, or, if the
current test is a test file, the (unique) beginning of a test block
name.
\item \code{".."}, in which case we step \emph{out}, i.e. from a test block to the
test file, from a test file to the test directory including all
tests.
\item \code{"/"}, to select the test directory including all tests.
}}
}
\description{
The current working test can be:
\itemize{
\item all tests in the test directory,
\item all tests in a test file,
\item all expectation in a single test block.
}
}
\details{
If \code{x} does not uniquely specify the test file or test block, and the
R session is interactive, then the user needs to select between them
manually. For non-interactive R sessions and error is thrown in this
case.

\code{setwt} (set working test) is a synomym of \code{ct} (change test).
}
\examples{
TODO
}
\seealso{
Other test selection functions: \code{\link{pwt}}
}
