% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current-test.R
\name{ct_get_mode}
\alias{ct_get_mode}
\title{Set or query the mocking status of the current test}
\usage{
ct_get_mode()
}
\value{
\code{ct_get_mode} returns \code{"mock"}, \code{"nomock"} or \code{"record"}.
}
\description{
Note that these settings only apply to the \emph{current test}. Other
test cases are always mocked.
}
\details{
\code{ct_get_mode} queries the current mode.

\code{ct_set_mode} sets the current mode.

Modes:
\itemize{
\item \code{nomock} turns off mocking. for the current test. You typically want
this while you are writing or extending the test cases. \code{httrmock} goes
completely out of the way, when this mode is selected.
\item record` turns on recording mode. All HTTP responses are recorded
in this mode.
\item \code{mock} turns on mocking mode. HTTP requests are not performed,
the recorded responses are replayed. This is the default state after
you have changed the current test.
}
}
