% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current-test.R
\name{ct_into_block}
\alias{ct_into_block}
\title{Step into a test block}
\usage{
ct_into_block(current, block)
}
\arguments{
\item{current}{Current test.}

\item{block}{Block name or beginning of the name
(possible empty string).}
}
\description{
Step into a test block
}
\keyword{internal}
