% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{ct_into_file}
\alias{ct_into_file}
\title{Step into a test file}
\usage{
ct_into_file(current, file)
}
\arguments{
\item{current}{Current test.}

\item{file}{File name or beginning of the name (without the \code{test[-_]?}
prefix), possibly empty string.}
}
\description{
Step into a test file
}
