% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{ct_mode}
\alias{ct_mode}
\alias{ct_nomock}
\alias{ct_record}
\alias{ct_mock}
\title{Set or query the mocking status of the current test}
\usage{
ct_mode()

ct_nomock()

ct_record()

ct_mock()
}
\value{
\code{ct_mode} returns \code{"mock"}, \code{"nomock"} or \code{"record"}.
}
\description{
Note that these settings only apply to the \emph{current test}. Other
test cases are always mocked.
}
\details{
\code{ct_mode} queries the current state.

\code{ct_nomock} turns off mocking. for the current test. You typically want
this while you are writing or extending the test cases. \code{httrmock} goes
completely out of the way, when this mode is selected.

\code{ct_record} turns on recording mode. All HTTP responses are recorded
in this mode.

\code{ct_mock} turns on mocking mode. HTTP requests are not performed,
the recorded responses are replayed. This is the default state after
you have changed the current test.
}
