% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{httrmock}
\alias{httrmock}
\alias{httrmock-package}
\title{Mock HTTP Requests for API Testing}
\description{
Mock HTTP Requests for API Testing
}
\section{Introduction}{


TODO

\code{httrmock} was inspired by https://github.com/assaf/node-replay
}

\section{Usage in packages}{


TODO
}

\section{The database}{


TODO
}

\section{Debugging}{


\code{httrmock} uses \code{debugme} for easy debugging of what is recorded and
replayed, see \code{\link[debugme:debugme]{debugme::debugme()}} for details.
}

\seealso{
Other HTTP mocking: \code{\link{clear_recordings}},
  \code{\link{list_recordings}}
}
