% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{list_recordings}
\alias{list_recordings}
\title{List all HTTP request and response recordings}
\usage{
list_recordings()
}
\value{
A data frame with columns:
\itemize{
\item \code{id} The id of the recording. You can use this to delete it with
\code{\link[=del_recording]{del_recording()}}.
\item \code{method} The HTTP method.
\item \code{url} The URL used in the request. Note that the URL in the response
might be different.
\item \code{user} The local username of the user that recorded the request,
as obtained via \code{\link[whoami:username]{whoami::username()}}.
\item \code{timestamp} The exact date and time when the request was recorded.
}
}
\description{
List all HTTP request and response recordings
}
\seealso{
Other HTTP mocking: \code{\link{clear_recordings}},
  \code{\link{httrmock}}, \code{\link{mocking_status}},
  \code{\link{stop_recording}},
  \code{\link{with_httrmock_context}}
}

