% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{mocking_status}
\alias{mocking_status}
\title{\code{httrmock} recording and replaying status}
\usage{
mocking_status()
}
\value{
A logical vector of two components: \sQuote{recording} and
\sQuote{replaying}.
}
\description{
\code{httrmock} recording and replaying status
}
\seealso{
Other HTTP mocking: \code{\link{clear_recordings}},
  \code{\link{httrmock}}, \code{\link{list_recordings}},
  \code{\link{stop_recording}}
}

