% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{pwt}
\alias{pwt}
\alias{getwt}
\title{Get the current working test}
\usage{
pwt()

getwt()
}
\value{
Path to a test direcory, test file, or path and name of the
test block.
}
\description{
See \code{\link[=ct]{ct()}} for a better description of current working tests.
}
\details{
\code{getwt} (get working test) is a synonym of \code{pwt} (print working test).
}
\examples{
TODO
}
\seealso{
Other test selection functions: \code{\link{ct}}
}
