% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{start_recording}
\alias{start_recording}
\title{Start or stop recording or replaying HTTP requests}
\usage{
start_recording()
}
\description{
\code{start_recording()} starts recording, and \code{stop_recording()} stops it.
}
\seealso{
Other HTTP mocking: \code{\link{clear_recordings}},
  \code{\link{httrmock}}, \code{\link{list_recordings}}
}
