% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{transform_request_for_recording}
\alias{transform_request_for_recording}
\title{Transform a Request for Recording}
\usage{
transform_request_for_recording(request)
}
\arguments{
\item{request}{The httr request object.}
}
\value{
The transformed request, that can be written to a file easily.
}
\description{
We will not use the result to (re)create the request, we only use it
for matching, so the transformation does not have to be lossless.
}
\keyword{internal}
