% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{httrmock}
\alias{httrmock}
\alias{httrmock-package}
\title{Mock HTTP Requests for API Testing}
\description{
Mock HTTP Requests for API Testing
}
\section{Introduction}{


TODO

\code{httrmock} was inspired by https://github.com/assaf/node-replay
}

\section{Usage in packages}{


TODO
}

\section{The database}{


\code{httrmock} uses the \code{storr} package to store the requests and responses
in RDS files, see \code{\link[storr:storr]{storr::storr()}}. By default the database is in the
\code{tests/testthat} directory, which is appropriate for \code{testthat} tests.
For an alternative directory, set the the \code{HTTRMOCK_STORE} environment
variable and point it to the directory you wish to use. The directory
will be created if it does not exist.
}

\section{Debugging}{


\code{httrmock} uses \code{debugme} for easy debugging of what is recorded and
replayed, see \code{\link[debugme:debugme]{debugme::debugme()}} for details.
}

\seealso{
Other HTTP mocking: \code{\link{clear_recordings}},
  \code{\link{list_recordings}},
  \code{\link{start_recording}}
}
