% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{stop_recording}
\alias{stop_recording}
\alias{stop_replaying}
\alias{start_recording}
\alias{start_replaying}
\title{Start or stop recording or replaying HTTP requests}
\usage{
stop_recording()

stop_replaying()

start_recording()

start_replaying()
}
\description{
\code{start_recording()} starts recording, and \code{stop_recording()} stops it.
}
\seealso{
Other HTTP mocking: \code{\link{clear_recordings}},
  \code{\link{httrmock}}, \code{\link{list_recordings}},
  \code{\link{mocking_status}},
  \code{\link{with_httrmock_context}}
}

